Chat overlay broken down into message and input component -> need to integrate functionality...
