+++ packages/desktop/src/components/chat/components/input/index.js
@@ -1,0 +1,3 @@
+import Input from './input';
+
+export default Input;
