+++ packages/desktop/src/components/chat/components/messages/index.js
@@ -1,0 +1,3 @@
+import Messages from './messages';
+
+export default Messages;
