+++ packages/desktop/src/components/shared/overlay/components/header/index.js
@@ -1,0 +1,3 @@
+import Header from './header';
+
+export default Header;
